@@ -6,11 +6,9 @@
   "main": "index.js",
   "scripts": {
     "dev:build-server": "webpack --config webpack.server.js --watch",
-    "dev:run-server":"nodemon --watch server-build --exec node server-build/bundle.js ",
-
-    "dev:build-client":"webpack --config webpack.client.js --watch",
-"start": "npm-run-all --parallel dev:*",
-
+    "dev:run-server": "nodemon --watch server-build --exec node server-build/bundle.js ",
+    "dev:build-client": "webpack --config webpack.client.js --watch",
+    "start": "npm-run-all --parallel dev:*",
     "test": "./node_modules/.bin/jest --watchAll",
     "start_old": "webpack-dev-server --debug --devtool cheap-module-source-map --output-pathinfo --history-api-fallback --hot --inline --progress --colors --open",
     "build_old": "webpack -p --progress --colors"
@@ -46,9 +44,9 @@
     "@types/react-redux": "5.0.13",
     "@types/react-router-dom": "^4.0.7",
     "@types/redux-actions": "2.2.2",
+    "@types/tapable": "0.2.4",
     "@types/webpack": "^3.0.0",
     "@types/webpack-env": "1.13.2",
-    "@types/tapable": "0.2.4",
     "awesome-typescript-loader": "^3.4.0",
     "css-loader": "^0.28.7",
     "enzyme": "^3.3.0",
@@ -69,16 +67,18 @@
     "react-hot-loader": "^3.1.3",
     "react-test-renderer": "^16.2.0",
     "redux-devtools-extension": "^2.13.2",
+    "serverless-offline": "^3.32.1",
+    "serverless-webpack": "^5.2.0",
     "style-loader": "^0.19.0",
     "ts-jest": "^22.0.1",
     "types": "^0.1.1",
     "typescript": "^2.6.1",
     "url-loader": "^0.6.2",
-    "webpack": "^3.8.1",
+    "webpack": "^3.12.0",
     "webpack-bundle-analyzer": "^2.10.0",
     "webpack-dev-server": "^2.9.4",
     "webpack-merge": "4.1.0",
-    "webpack-node-externals": "1.6.0"
+    "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
     "@types/express": "^4.16.0",
@@ -101,6 +101,7 @@
     "redux-persist": "^5.7.0",
     "redux-thunk": "^2.2.0",
     "serialize-javascript": "1.4.0",
+    "serverless-http": "^1.8.0",
     "styled-components": "^3.2.3"
   }
 }
