@@ -1,11 +1,13 @@
 import * as express from 'express';
 import * as React from 'react';
 import * as ReactDOMServer from 'react-dom/server';
+import serverless from "serverless-http";
+
 import { Home } from './client/components/Home';
 const app = express();
 
 app.use(express.static('public'));
-app.get('/', (req, res) => {
+app.get('/one', (req, res) => {
     const content = ReactDOMServer.renderToString(<Home />);
     const html=`<html>
     <head>
@@ -26,4 +28,5 @@ app.listen(3000, () => {
     console.log("listining on port 3000");
 })
 
-module.exports = app;
+export const handler=serverless(app);
+//module.exports = app;
