@@ -73,30 +73,24 @@ module.exports = require("react");
 /* 1 */
 /***/ (function(module, exports, __webpack_require__) {
 
-__webpack_require__(2);
-module.exports = __webpack_require__(3);
+module.exports = __webpack_require__(2);
 
 
 /***/ }),
 /* 2 */
-/***/ (function(module, exports) {
-
-module.exports = require("babel-polyfill");
-
-/***/ }),
-/* 3 */
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
 
 Object.defineProperty(exports, "__esModule", { value: true });
-var express = __webpack_require__(4);
+var express = __webpack_require__(3);
 var React = __webpack_require__(0);
-var ReactDOMServer = __webpack_require__(5);
+var ReactDOMServer = __webpack_require__(4);
+var serverless_http_1 = __webpack_require__(5);
 var Home_1 = __webpack_require__(6);
 var app = express();
 app.use(express.static('public'));
-app.get('/', function (req, res) {
+app.get('/one', function (req, res) {
     var content = ReactDOMServer.renderToString(React.createElement(Home_1.Home, null));
     var html = "<html>\n    <head>\n\n    <link rel=\"stylesheet\" href=\"https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css\">\n\n        </head>\n<body>\n<div id=\"root\">" + content + "</div>\n<script src=\"bundle.js\"></script>\n</body>\n    </html>\n    \n    ";
     res.send(html);
@@ -104,22 +98,28 @@ app.get('/', function (req, res) {
 app.listen(3000, function () {
     console.log("listining on port 3000");
 });
-module.exports = app;
+exports.handler = serverless_http_1.default(app);
 
 
  ;(function register() { /* react-hot-loader/webpack */ if (process.env.NODE_ENV !== 'production') { if (typeof __REACT_HOT_LOADER__ === 'undefined') { return; } /* eslint-disable camelcase, no-undef */ var webpackExports = typeof __webpack_exports__ !== 'undefined' ? __webpack_exports__ : module.exports; /* eslint-enable camelcase, no-undef */ if (typeof webpackExports === 'function') { __REACT_HOT_LOADER__.register(webpackExports, 'module.exports', "C:\\JSPOC\\serverless\\own-sample\\src\\index.tsx"); return; } /* eslint-disable no-restricted-syntax */ for (var key in webpackExports) { /* eslint-enable no-restricted-syntax */ if (!Object.prototype.hasOwnProperty.call(webpackExports, key)) { continue; } var namedExport = void 0; try { namedExport = webpackExports[key]; } catch (err) { continue; } __REACT_HOT_LOADER__.register(namedExport, key, "C:\\JSPOC\\serverless\\own-sample\\src\\index.tsx"); } } })();
 
 /***/ }),
-/* 4 */
+/* 3 */
 /***/ (function(module, exports) {
 
 module.exports = require("express");
 
 /***/ }),
-/* 5 */
+/* 4 */
 /***/ (function(module, exports) {
 
 module.exports = require("react-dom/server");
+
+/***/ }),
+/* 5 */
+/***/ (function(module, exports) {
+
+module.exports = require("serverless-http");
 
 /***/ }),
 /* 6 */
